requirements.txt
tensorflow-macos==2.12.0
tensorflow-metal
pandas
numpy
dvc
notebook
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
scipy
Flask
Flask-Cors
-e .